+++ zfit/minimizers/minimizer_nlopt.py
@@ -188,7 +188,7 @@ class NLoptBaseMinimizerV1(BaseMinimizer):
         minimizer = nlopt.opt(nlopt.LD_LBFGS, len(params))
 
         # initial values as array
-        xvalues = np.array(run(params))
+        xvalues = initial_xvalues = np.asarray(run(params))
 
         # get and set the limits
         lower = np.array([p.lower for p in params])
@@ -201,11 +201,11 @@ class NLoptBaseMinimizerV1(BaseMinimizer):
         def obj_func(x, grad):
             if grad.size > 0:
                 value, gradients = evaluator.value_gradient(x)
-                grad[:] = np.array(run(gradients))
+                grad[:] = np.asarray(gradients)
             else:
                 value = evaluator.value(x)
 
-            return value
+            return float(value)
 
         minimizer.set_min_objective(obj_func)
 
@@ -268,7 +268,6 @@ class NLoptBaseMinimizerV1(BaseMinimizer):
                 if step_size is None and param.has_step_size:
                     step_size = param.step_size
                 init_scale.append(step_size)
-
             minimizer.set_initial_step(init_scale)
 
             self._set_tols_inplace(
@@ -300,17 +299,27 @@ class NLoptBaseMinimizerV1(BaseMinimizer):
                         "Minimization in NLopt failed, restarting with slightly varied parameters."
                     )
                 if nrandom < self._nrandom_max:  # in order not to start too close
+                    init_scale_isnot_none = np.asarray(
+                        [scale is not None for scale in init_scale], dtype=np.bool
+                    )
                     init_scale = np.where(
-                        init_scale != None, init_scale, np.ones_like(init_scale)
+                        init_scale_isnot_none,
+                        init_scale,
+                        np.ones_like(init_scale, dtype=np.float64),
                     )
-                    init_scale_no_nan = np.nan_to_num(init_scale, nan=1.0)
 
-                    xvalues += (
-                        np.random.uniform(
-                            low=-init_scale_no_nan, high=init_scale_no_nan
-                        )
-                        / 2
+                    init_scale_no_nan = np.nan_to_num(init_scale, nan=1.0)
+                    init_scale_no_nan = init_scale_no_nan.astype(np.float64)
+                    upper_random = np.minimum(
+                        initial_xvalues + init_scale_no_nan / 2, upper
                     )
+                    lower_random = np.maximum(
+                        initial_xvalues - init_scale_no_nan / 2, lower
+                    )
+                    initial_xvalues = np.random.uniform(
+                        low=lower_random, high=upper_random
+                    )
+
                     nrandom += 1
             else:
                 maxiter_reached = evaluator.niter > evaluator.maxiter
@@ -1425,7 +1434,7 @@ class NLoptMLSLV1(NLoptBaseMinimizerV1):
             algorithm = nlopt.GD_MLSL_LDS
 
         local_minimizer = nlopt.LD_LBFGS if local_minimizer is None else local_minimizer
-        if not isinstance(local_minimizer, collections.Mapping):
+        if not isinstance(local_minimizer, collections.abc.Mapping):
             local_minimizer = {"algorithm": local_minimizer}
         if "algorithm" not in local_minimizer:
             raise ValueError("algorithm needs to be specified in local_minimizer")